+++ govcd/lbservicemonitor_test.go
@@ -12,7 +12,7 @@ import (
 )
 
 // Test_LBServiceMonitor tests CRUD methods for load balancer service monitor.
-// The following things are tested:
+// The following things are tested if prerequisite Edge Gateway exists:
 // Creation of load balancer service monitor
 // Read load balancer by both ID and Name (service monitor name must be unique in single edge gateway)
 // Update - change a single field and compare that configuration and result objects are deeply equal
